various_template_changes Geometry|vector|circle|MODINT

+++ various/default.cpp
@@ -32,7 +32,7 @@ template<typename A, typename B> ostream& operator<<(ostream &cout, pair<A, B> c
 template<typename A> ostream& operator<<(ostream &cout,vector<A> const &v){cout<<"[";for(int i=0;i<v.size();i++){if(i)cout<<", ";cout<<v[i];}return cout<<"]";}
 template<typename A, typename B> istream& operator>>(istream& cin, pair<A, B> &p){cin>>p.F;return cin>>p.S;}
 
-const size_t MAXN = 1e5 +7;
+const long long MAXN = 1e5 +7;
 
 void check(){
